split up animation loader and bone animator
